# [PERF] product_product: avoid injecting ids in product name_search