--- a/addons/product/models/product_product.py
+++ b/addons/product/models/product_product.py
@@ -509,11 +509,12 @@ class ProductProduct(models.Model):
                 # on a database with thousands of matching products, due to the huge merge+unique needed for the
                 # OR operator (and given the fact that the 'name' lookup results come from the ir.translation table
                 # Performing a quick memory merge of ids in Python will give much better performance
-                product_ids = list(self._search(args + [('default_code', operator, name)], limit=limit))
+                products_query = self._search(args + [('default_code', operator, name)], limit=limit)
+                product_ids = list(products_query)
                 if not limit or len(product_ids) < limit:
                     # we may underrun the limit because of dupes in the results, that's fine
                     limit2 = (limit - len(product_ids)) if limit else False
-                    product2_ids = self._search(args + [('name', operator, name), ('id', 'not in', product_ids)], limit=limit2, access_rights_uid=name_get_uid)
+                    product2_ids = self._search(args + [('name', operator, name), ('id', 'not in', products_query)], limit=limit2, access_rights_uid=name_get_uid)
                     product_ids.extend(product2_ids)
             elif not product_ids and operator in expression.NEGATIVE_TERM_OPERATORS:
                 domain = expression.OR([