Before this PR, the name_search function would perform a search on products by injecting a list of ids in the domain instead of using a subquery. In cases where this list of ids is way too big, the search query becomes extremely slow.

This PR uses a subquery instead in the domain to avoid this problem.

Benchmarks:

|Num. product_ids| Before PR | After PR |
|---------------------|---------------|--------------| 
|28801| 65 s| <1 s|

opw-4743566

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr